% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_alleles.R
\name{calculate_alleles}
\alias{calculate_alleles}
\title{Calculate Allele length}
\usage{
calculate_alleles(
  repeat_counts,
  scale = TRUE,
  center = TRUE,
  plot = FALSE,
  z.norm.thresh = 3
)
}
\arguments{
\item{repeat_counts}{a numeric or integer vector corresponding to repeats}

\item{scale}{if the data should be scaled; reccommended do not change}

\item{center}{if the data should be centered; reccommended do not change}

\item{plot}{provide some helpful plots to visualize what function is doing}

\item{z.norm.thresh}{the zscore threshold used to remove outliers in data. NOTE: long tails at end of expanded allele will skew expanded allele calling; recommended to remove outliers}
}
\value{
A dataframe containing the insert, barcode ID, sample ID, and direction of the insert
}
\description{
This function was designed to estimate the length of both the 'normal' inherited allele and the inherited expanded allele specifically in
CAG repeat disorders such as Huntingtons disease - but can be adapted for use in other expansion disorders.
This fuction expects a dataframe with
}
